requirements.txt
# Python Requirements untuk System Pakar Engine
# Rekomendasi Python: 3.11 atau 3.12 (kompatibel dengan Laravel 11 dan modern libraries)

# FastAPI untuk API server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# HTTP Client untuk komunikasi dengan Laravel API
requests>=2.31.0
aiohttp>=3.9.0

# Expert System - Experta untuk rule-based reasoning
experta>=1.9.0

# Scientific computing untuk perhitungan CF
numpy>=1.24.0

# WhatsApp API Client - Fonte
# Catatan: Sesuaikan dengan library Fonte yang akan digunakan
# Contoh: fonte-api atau library WhatsApp lainnya
# fonte-api>=1.0.0

# Environment variables
python-dotenv>=1.0.0

# Data validation
pydantic>=2.5.0

# Logging
loguru>=0.7.2

# Date/time handling
python-dateutil>=2.8.2

# CORS support untuk FastAPI
python-multipart>=0.0.6

